+++ bin/luarocks-tag-release-action.lua
@@ -87,8 +87,12 @@ local function parse_copy_directory_args(str_args)
   return filter_existing_directories(copy_directories)
 end
 
+local is_pull_request = getenv_or_empty('GITHUB_EVENT_NAME') == 'pull_request'
+
 local license_input = os.getenv('INPUT_LICENSE')
 local template_input = os.getenv('INPUT_TEMPLATE')
+local package_name = getenv_or_err('INPUT_NAME')
+local package_version = is_pull_request and '0.0.0' or getenv_or_err('INPUT_VERSION')
 
 local interpreters_input = os.getenv('INPUT_TEST_INTERPRETERS')
 ---@type lua_interpreter[]
@@ -112,8 +116,6 @@ local args = {
   github_repo = github_repo,
   repo_name = repo_name,
   github_server_url = github_server_url,
-  package_name = getenv_or_err('INPUT_NAME'),
-  package_version = getenv_or_err('INPUT_VERSION'),
   dependencies = parse_list_args(getenv_or_empty('INPUT_DEPENDENCIES')),
   labels = parse_list_args(getenv_or_empty('INPUT_LABELS')),
   copy_directories = parse_copy_directory_args(getenv_or_err('INPUT_COPY_DIRECTORIES')),
@@ -121,7 +123,7 @@ local args = {
   detailed_description_lines = parse_list_args(getenv_or_empty('INPUT_DETAILED_DESCRIPTION')),
   rockspec_template_file_path = template_input ~= '' and template_input
     or action_path .. '/resources/rockspec.template',
-  upload = getenv_or_err('INPUT_UPLOAD') == 'true',
+  upload = not is_pull_request,
   license = license_input ~= '' and license_input or nil,
   luarocks_test_interpreters = test_interpreters,
 }
@@ -130,10 +132,19 @@ table.insert(args.dependencies, 1, 'lua >= 5.1')
 args.ref_type = getenv_or_err('GITHUB_REF_TYPE')
 print('Workflow has been triggered by: ' .. args.ref_type)
 args.git_ref = getenv_or_err('GITHUB_REF_NAME')
-if args.ref_type ~= 'tag' then
+local is_tag = args.ref_type ~= 'tag'
+if not is_tag then
+  print('Publishing an untagged release.')
   args.git_ref = getenv_or_err('GITHUB_SHA')
 end
 
 local luarocks_tag_release = require('luarocks-tag-release')
 
-luarocks_tag_release(args)
+local specrev = '1'
+if is_pull_request then
+  print('Running in a pull request.')
+  specrev = assert(os.getenv('GITHUB_RUN_ATTEMPT'), 'GITHUB_RUN_ATTEMPT not set')
+  args.git_ref = getenv_or_err('GITHUB_SHA')
+end
+
+luarocks_tag_release(package_name, package_version, specrev, args)