+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Ability to test a local `luarocks install`, without uploading to luarocks.org
+  on `pull_request`.
+
 ## [v5.0.1] - 2023-07-06
 ### Fixed
 - Run `luarocks test` only if a `.busted` file exists in the project root.